Extract review actions into src/core/review-actions.ts
Move approve/deny/reopen/cancel, batch review, auto-deny shadowban, and
moderator removal actions into a dedicated module. Typecheck, 167 tests,
and build all pass.
